requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
pettingzoo>=1.24.0
gymnasium>=0.29.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0
matplotlib>=3.7.0
pandas>=2.0.0
